WIP upd

+++ src/config.hpp
@@ -2,6 +2,6 @@
 
 #include <cinttypes>
 
-constexpr uint8_t TILE_SIZE = 32;
-constexpr uint8_t TILES_IN_WORLD_CELL_W = 32;
-constexpr uint8_t TILES_IN_WORLD_CELL_H = 32;
+constexpr int TILE_SIZE = 32;
+constexpr int TILES_IN_WORLD_CELL_W = 32;
+constexpr int TILES_IN_WORLD_CELL_H = 32;
